Refactor item control actions to use common document sheet mixin handlers
