+++ module/applications/actor/common-sheet.mjs
@@ -1,5 +1,3 @@
-import ED4E from "../../config/_module.mjs";
-import { getSetting } from "../../settings.mjs";
 import DocumentSheetMixinEd from "../api/document-sheet-mixin.mjs";
 
 const { ActorSheetV2 } = foundry.applications.sheets;
@@ -16,27 +14,9 @@ export default class ActorSheetEd extends DocumentSheetMixinEd( ActorSheetV2 ) {
    * @userFunction UF_ActorSheetEd-defaultOptions
    */
   static DEFAULT_OPTIONS = {
-    classes:  [ "earthdawn4e", "sheet", "actor" ],
-    window:   {
-      frame:          true,
-      positioned:     true,
-      icon:           false,
-      minimizable:    true,
-      resizable:      true,
-    },
+    classes:  [ "actor", ],
     actions:  {
-      editImage:          ActorSheetEd._onEditImage,
-      editItem:           ActorSheetEd._onItemEdit,
-      deleteItem:         ActorSheetEd._onItemDelete,
-      displayItem:        ActorSheetEd._onDisplayItem,
       expandItem:         ActorSheetEd._onCardExpand,
-      createChild:        ActorSheetEd._onCreateChild,
-      deleteChild:        ActorSheetEd._onDeleteChild,
-      displayChildToChat: ActorSheetEd._onDisplayChildToChat,
-      editChild:          ActorSheetEd._onEditChild,
-    },
-    form: {
-      submitOnChange: true,
     },
   };
 
@@ -90,33 +70,18 @@ export default class ActorSheetEd extends DocumentSheetMixinEd( ActorSheetV2 ) {
     );
   }
 
-  /* -------------------------------------------- */
-  /*  Rendering                                   */
-  /* -------------------------------------------- */
+  // region Rendering
 
   /** 
    * @inheritdoc 
    * @userFunction UF_ActorSheetEd-prepareContext
    */
   async _prepareContext( options ) {
-    // TODO: überprüfen was davon benötigt wird
     const context = await super._prepareContext( options );
     foundry.utils.mergeObject( context, {
       actor:                  this.document,
-      system:                 this.document.system,
       items:                  this.document.items,
-      options:                this.options,
-      systemFields:           this.document.system.schema.fields,
-      config:                 ED4E,
     } );
-
-    context.enrichedDescription = await TextEditor.enrichHTML(
-      this.document.system.description.value,
-      {
-        // Only show secret blocks to owner
-        secrets:    this.document.isOwner,
-      }
-    );
 
     return context;
   }
@@ -129,68 +94,9 @@ export default class ActorSheetEd extends DocumentSheetMixinEd( ActorSheetV2 ) {
     return super._renderHTML( context, options );
   }
 
-  /* -------------------------------------------- */
-  /*  Drag and Drop                               */
-  /* -------------------------------------------- */
+  // endregion
 
-  /* -------------------------------------------- */
-  /*  Event Handlers                              */
-  /* -------------------------------------------- */
-
-  /**
-   * Changing the actor image
-   * @param {Event} event - The event that triggered the form submission.
-   * @param {HTMLElement} target - The HTML element that triggered the action.
-   * @returns {Promise<foundry.FilePicker>} - A promise that resolves when the image is changed.
-   * @userFunction UF_ActorSheetEd-onEditImage
-   */
-  static async _onEditImage( event, target ) {
-    const attr = target.dataset.edit;
-    const current = foundry.utils.getProperty( this.document, attr );
-    const { img } = this.document.constructor.getDefaultArtwork?.( this.document.toObject() ) ?? {};
-    // eslint-disable-next-line no-undef
-    const fp = new FilePicker( {
-      current,
-      type:           "image",
-      redirectToRoot: img ? [ img ] : [],
-      callback:       ( path ) => {
-        this.document.update( { [attr]: path } );
-      },
-      top:  this.position.top + 40,
-      left: this.position.left + 10,
-    } );
-    return fp.browse();
-  }
-
-  /**
-   * Editing an item
-   * @param {Event} event - The event that triggered the form submission.
-   * @param {HTMLElement} target - The HTML element that triggered the action.
-   * @returns {Promise<any>} - A promise that resolves when the item is edited.
-   * @userFunction UF_ActorSheetEd-onItemEdit
-   */
-  static async _onItemEdit( event, target ) {
-    event.preventDefault();
-    const itemId = target.parentElement.dataset.itemId;
-    const item = this.document.items.get( itemId );
-    return item.sheet?.render( true );
-  }
-
-  /**
-   * Deleting an item
-   * @param {Event} event - The event that triggered the form submission.
-   * @param {HTMLElement} target - The HTML element that triggered the action.
-   * @returns {Promise<void>} - A promise that resolves when the item is deleted.
-   * @userFunction UF_ActorSheetEd-onItemDelete
-   */
-  static async _onItemDelete( event, target ) {
-    event.preventDefault();
-    const itemId = target.parentElement.dataset.itemId;
-    const item = this.document.items.get( itemId );
-    if ( !item ) return;
-    if ( getSetting( "quickDeleteEmbeddedOnShiftClick" ) && event.shiftKey ) return item.delete();
-    else item.deleteDialog();
-  }
+  // region Event Handlers
 
   /**
    * Expanding or collapsing the item description
@@ -210,101 +116,6 @@ export default class ActorSheetEd extends DocumentSheetMixinEd( ActorSheetV2 ) {
     itemDescription.toggleClass( "card__description--toggle" );
   }
 
-  /**
-   * Display an item in the chat
-   * @param {Event} event - The event that triggered the form submission.
-   * @param {HTMLElement} target - The HTML element that triggered the action.
-   * @returns {Promise<void>} - A promise that resolves when the item is displayed in the chat.
-   * @userFunction UF_ActorSheetEd-onDisplayItem
-   */
-  static async _onDisplayItem( event, target ) {
-    ui.notifications.info( "Display Item not done yet" );
-  }
-
-  /** 
-   * Creates a new child active effect
-   * @param {Event} event - The event that triggered the form submission.
-   * @param {HTMLElement} target - The HTML element that triggered the action.
-   * @returns {Promise<foundry.abstract.Document>} - A promise that resolves when the child is created.
-   * @userFunction UF_ActorSheetEd-onCreateChild
-   */
-  static async _onCreateChild( event, target ) {
-    const type = target.dataset.type;
-
-    if ( type === "effect" ) return ActiveEffect.implementation.create( {
-      type:     "eae",
-      name:     game.i18n.localize( "ED.ActiveEffect.newEffectName" ),
-      icon:     "icons/svg/aura.svg",
-      changes:  [ {} ],
-      system:  {
-        duration: {
-          type: target.dataset.effectPermanent ? "permanent" : "combat",
-        },
-        changes: [ {} ],
-      },
-    }, {
-      parent:      this.document,
-      renderSheet: true,
-    } );
-
-    // this will make more sense when we have a common documentSheet mixin
-    /* if ( activeTab === "spells" ) return Item.implementation.create({
-      name: game.i18n.format("DOCUMENT.New", { type: game.i18n.format(CONFIG.Item.typeLabels.spell) }),
-      type: "spell",
-      img: Item.implementation.getDefaultArtwork({ type: "spell" })?.img ?? Item.implementation.DEFAULT_ICON
-    }, { parent: this.actor, renderSheet: true });
-
-    const features = ["feat", "race", "background", "class", "subclass"];
-    if ( this.actor.type === "npc" ) features.push("weapon");
-
-    let types = {
-      features,
-      inventory: ["weapon", "equipment", "consumable", "tool", "container", "loot"]
-    }[activeTab] ?? [];
-
-    types = types.filter(type => {
-      const model = CONFIG.Item.dataModels[type];
-      return !model.metadata?.singleton || !this.actor.itemTypes[type].length;
-    });
-
-    if ( types.length ) return Item.implementation.createDialog({}, {
-      parent: this.actor, pack: this.actor.pack, types
-    }); */
-  }
-
-  /** 
-   * Deletes a child active effect
-   * @param {Event} event - The event that triggered the form submission.
-   * @param {HTMLElement} target - The HTML element that triggered the action.
-   * @returns {Promise<foundry.abstract.Document>} - A promise that resolves when the child is deleted.
-   * @userFunction UF_ActorSheetEd-onDeleteChild
-   */
-  static async _onDeleteChild( event, target ) {
-    const document = await fromUuid( target.dataset.uuid );
-    if ( getSetting( "quickDeleteEmbeddedOnShiftClick" ) && event.shiftKey ) return document.delete();
-    else document.deleteDialog();
-  }
-
-  /** 
-   * Displays a child active effect in the chat
-   * @param {Event} event - The event that triggered the form submission.
-   * @param {HTMLElement} target - The HTML element that triggered the action.
-   * @returns {Promise<foundry.abstract.Document>} - A promise that resolves when the child is displayed in chat.
-   * @userFunction UF_ActorSheetEd-onDisplayChildToChat
-   */
-  static async _onDisplayChildToChat( event, target ) {
-    ChatMessage.create( { content: "Coming up: a beautiful description of the Item you just clicked to be displayed here in chat!" } );
-  }
-
-  /** 
-   * Edits a child active effect
-   * @param {Event} event - The event that triggered the form submission.
-   * @param {HTMLElement} target - The HTML element that triggered the action.
-   * @returns {Promise<foundry.abstract.Document>} - A promise that resolves when the child is edited.
-   * @userFunction UF_ActorSheetEd-onEditChild
-   */
-  static async _onEditChild( event, target ) {
-    ( await fromUuid( target.dataset.uuid ) ).sheet?.render( { force: true } );
-  }
+  // endregion
 
 }
