+++ module/applications/item/item-sheet.mjs
@@ -1,9 +1,9 @@
-import ED4E from "../../config/_module.mjs";
 import SpellEnhancementsConfig from "../configs/spell-enhancements-config.mjs";
 import ConstraintsConfig from "../configs/constraints-config.mjs";
 import DocumentSheetMixinEd from "../api/document-sheet-mixin.mjs";
 
 const { ItemSheetV2 } = foundry.applications.sheets;
+const { DragDrop, TextEditor } = foundry.applications.ux;
 
 // noinspection JSClosureCompilerSyntax
 /**
@@ -17,24 +17,9 @@ export default class ItemSheetEd extends DocumentSheetMixinEd( ItemSheetV2 ) {
   static DEFAULT_OPTIONS = {
     id:       "item-sheet-{id}",
     uniqueId: String( ++foundry.applications.api.ApplicationV2._appId ),
-    classes:  [ "earthdawn4e", "sheet", "item" ],
-    window:   {
-      frame:          true,
-      positioned:     true,
-      icon:           false,
-      minimizable:    true,
-      resizable:      true,
-    },
-    form: {
-      submitOnChange: true,
-    },
+    classes:  [ "item", ],
     actions:  {
       config:             ItemSheetEd._onConfig,
-      editImage:          ItemSheetEd._onEditImage,
-      createChild:        ItemSheetEd._onCreateChild,
-      deleteChild:        ItemSheetEd._onDeleteChild,
-      displayChildToChat: ItemSheetEd._onDisplayChildToChat,
-      editChild:          ItemSheetEd._onEditChild,
     },
     position: {
       top:    50,
@@ -117,20 +102,7 @@ export default class ItemSheetEd extends DocumentSheetMixinEd( ItemSheetV2 ) {
       context,
       {
         item:                   this.document,
-        system:                 this.document.system,
         options:                this.options,
-        systemFields:           this.document.system.schema.fields,
-        isGM:                   game.user.isGM,
-        config:                 ED4E,
-      }
-    );
-
-    context.enrichedDescription = await TextEditor.enrichHTML(
-      this.document.system.description.value,
-      {
-        // Only show secret blocks to owner
-        secrets:    this.document.isOwner,
-        EdRollData: this.document.getRollData
       }
     );
 
@@ -187,83 +159,6 @@ export default class ItemSheetEd extends DocumentSheetMixinEd( ItemSheetV2 ) {
     app?.render( { force: true } );
   }
 
-  static async _onEditImage( event, target ) {
-    const attr = target.dataset.edit;
-    const current = foundry.utils.getProperty( this.document, attr );
-    const { img } = this.document.constructor.getDefaultArtwork?.( this.document.toObject() ) ?? {};
-    // eslint-disable-next-line no-undef
-    const fp = new FilePicker( {
-      current,
-      type:           "image",
-      redirectToRoot: img ? [ img ] : [],
-      callback:       ( path ) => {
-        this.document.update( { [attr]: path } );
-      },
-      top:  this.position.top + 40,
-      left: this.position.left + 10,
-    } );
-    return fp.browse();
-  }
-
-  /** @inheritDoc */
-  static async _onCreateChild( event, target ) {
-    const type = target.dataset.type;
-
-    if ( type === "effect" ) return ActiveEffect.implementation.create( {
-      type:     "eae",
-      name:     game.i18n.localize( "ED.ActiveEffect.newEffectName" ),
-      icon:     "icons/svg/aura.svg",
-      changes:  [ {} ],
-      duration: {
-        permanent: !!target.dataset.effectPermanent,
-      },
-      system:  {
-        changes: [ {} ],
-      },
-    }, {
-      parent:      this.document,
-      renderSheet: true,
-    } );
-
-    // this will make more sense when we have a common documentSheet mixin
-    /* if ( activeTab === "spells" ) return Item.implementation.create({
-      name: game.i18n.format("DOCUMENT.New", { type: game.i18n.format(CONFIG.Item.typeLabels.spell) }),
-      type: "spell",
-      img: Item.implementation.getDefaultArtwork({ type: "spell" })?.img ?? Item.implementation.DEFAULT_ICON
-    }, { parent: this.actor, renderSheet: true });
-
-    const features = ["feat", "race", "background", "class", "subclass"];
-    if ( this.actor.type === "npc" ) features.push("weapon");
-
-    let types = {
-      features,
-      inventory: ["weapon", "equipment", "consumable", "tool", "container", "loot"]
-    }[activeTab] ?? [];
-
-    types = types.filter(type => {
-      const model = CONFIG.Item.dataModels[type];
-      return !model.metadata?.singleton || !this.actor.itemTypes[type].length;
-    });
-
-    if ( types.length ) return Item.implementation.createDialog({}, {
-      parent: this.actor, pack: this.actor.pack, types
-    }); */
-  }
-
-  /** @inheritDoc */
-  static async _onDeleteChild( event, target ) {
-    ( await fromUuid( target.dataset.uuid ) ).delete();
-  }
-
-  /** @inheritDoc */
-  static async _onDisplayChildToChat( event, target ) {
-    ChatMessage.create( { content: "Coming up: a beautiful description of the Item you just clicked to be displayed here in chat!" } );
-  }
-
-  /** @inheritDoc */
-  static async _onEditChild( event, target ) {
-    ( await fromUuid( target.dataset.uuid ) ).sheet?.render( { force: true } );
-  }
   // endregion
 
   // region Drag and Drop
@@ -373,7 +268,7 @@ export default class ItemSheetEd extends DocumentSheetMixinEd( ItemSheetV2 ) {
    * @protected
    */
   async _onDropItem( event, item ) {
-    if ( !this.item.isOwner ) return;
+    // do nothing
   }
 
   /**
