+++ templates/actor/cards/ability-card.hbs
@@ -10,7 +10,7 @@
                 data-action="rollable" src="icons/svg/d20-black.svg">
             {{else}} 
             {{#if (ne this.system.strain 0)}}
-                <img class="ability-icon" src="{{img}}" data-edit="img" title="{{name}}" height="25" width="25" />
+                <img class="ability-icon" src="{{img}}"  data-edit="img" title="{{name}}" height="25" width="25" />
                 <img class="roll-icon__hover-fixed blood" src="icons/svg/blood.svg">
                 <img class="roll-icon__pointer take-strain" data-action="takeStrain" src="systems/ed4e/assets/icons/red-blood.svg" data-attribute="{{@key}}">
             {{else}}
