+++ templates/global/card-options-chat.hbs
@@ -1,11 +1,11 @@
 <div class="itemControlGroup" data-item-id="{{_id}}">
-  <a class="item-control item-display" data-action="displayItem" title="{{localize "ED.Actor.Cards.chat"}}">
+  <a class="item-control item-display" data-action="displayChild" title="{{localize "ED.Actor.Cards.chat"}}">
     <i class="fas fa-comments"></i>
   </a>
-  <a class="item-control item-delete" data-action="deleteItem" title="{{localize "ED.Actor.Cards.delete"}}">
+  <a class="item-control item-delete" data-action="deleteChild" data-uuid="{{uuid}}" title="{{localize "ED.Actor.Cards.delete"}}">
     <i class="fas fa-trash"></i>
   </a>
-  <a class="item-control item-edit" data-action="editItem" title="{{localize "ED.Actor.Cards.edit"}}">
+  <a class="item-control item-edit" data-action="editChild" data-uuid="{{uuid}}" title="{{localize "ED.Actor.Cards.edit"}}">
     <i class="fas fa-edit"></i>
   </a>
 </div>
