+++ templates/global/card-options-class-upgrade.hbs
@@ -2,16 +2,16 @@
   <a class="item-control item-upgrade__class" data-action="upgradeItem" title="{{localize 'earthdawn.u.upgrade'}}">
     <i class="fas fa-arrow-circle-up"></i>
   </a>
-  {{!-- <a class="item-control item-downgrade" data-action="displayItem" title="{{localize 'earthdawn.d.downgrade'}}">
+  {{!-- <a class="item-control item-downgrade" data-action="displayChild" title="{{localize 'earthdawn.d.downgrade'}}">
     <i class="fas fa-arrow-circle-down"></i>
   </a> --}}
-  <a class="item-control item-display" data-action="displayItem" title="Display Item">
+  <a class="item-control item-display" data-action="displayChild" title="Display Item">
     <i class="fas fa-comments"></i>
   </a>
-  <a class="item-control item-delete" data-action="deleteItem" title="Delete Item">
+  <a class="item-control item-delete" data-action="deleteChild" data-uuid="{{uuid}}" title="{{localize "ED.Actor.Cards.delete"}}">
     <i class="fas fa-trash"></i>
   </a>
-  <a class="item-control item-edit" data-action="editItem" title="Edit Item">
+  <a class="item-control item-edit" data-action="editChild" data-uuid="{{uuid}}" title="{{localize "ED.Actor.Cards.edit"}}">
     <i class="fas fa-edit"></i>
   </a>
 </div>
