+++ templates/global/card-options-effect.hbs
@@ -1,5 +1,5 @@
 <div class="itemControlGroup" data-item-id="{{_id}}">
-  <a class="item-control item-display" data-action="displayChildToChat" data-uuid="{{uuid}}" title="{{localize "ED.Actor.Cards.chat"}}">
+  <a class="item-control item-display" data-action="displayChild" data-uuid="{{uuid}}" title="{{localize "ED.Actor.Cards.chat"}}">
     <i class="fas fa-comments"></i>
   </a>
   <a class="item-control effect-delete" data-action="deleteChild" data-uuid="{{uuid}}" title="{{localize "ED.Actor.Cards.delete"}}">
