+++ templates/item/item-partials/item-section-top.hbs
@@ -1,6 +1,6 @@
 <div>
     <div class="partial__two">
-        <img src="{{item.img}}" data-edit="img" />
+        <img class="item-image" src="{{item.img}}" title="{{item.name}}" data-action="editImage" data-edit="img" />
     </div>
